occ-daily-volume/requirements.txt
argparse>=1.4.0
pandas>=2.0.0,<3.0.0
pyyaml>=6.0.0,<7.0.0
python-dateutil>=2.8.0,<3.0.0
requests>=2.31.0,<3.0.0
tabulate>=0.9.0,<1.0.0
